Clean up StoreCopy output formatting
The StoreCopy tool's output formatting doesn't behave reliably when
there's a lot of compaction going on.  A lot of the '.'s are skipped and
the line-ending summaries are skipped when a node or relationship
with the current node ID or rel ID doesn't exist (because that
information is outputted after getting the node / relationship).  This
resolves that so that every line is approximately the same length and
there's a line per every 500000 nodes/relationships.

This change also ups the number of nodes per '.' and per line.